add and update user orm and service

+++ src/main/kotlin/me/edurevsky/blog/blogqualquer/repositories/RoleRepository.kt
@@ -3,4 +3,4 @@ package me.edurevsky.blog.blogqualquer.repositories
 import me.edurevsky.blog.blogqualquer.entities.Role
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface RoleRepository : JpaRepository<Role, Long>
+interface RoleRepository : JpaRepository<Role, Int>
